services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.0
gotrue==2.8.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.0
httpx==0.25.2
bcrypt==4.2.1
docxtpl==0.16.8
python-docx==1.1.0
Pillow==10.3.0
qrcode==7.4.2
reportlab==4.2.2
